backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.1.1
groq==0.11.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
torch>=2.6.0
numpy>=1.26.4
